remove polling remote execution client
