+++ src/rust/engine/process_executor/src/main.rs
@@ -198,14 +198,6 @@ async fn main() {
                         This will likely start a new Nailgun server as a side effect.")
       )
       .arg(
-        Arg::with_name("enable-streaming-client")
-            .long("enable-streaming-client")
-            .takes_value(false)
-            .required(false)
-            .default_value("false")
-            .help("Enable the experimental streaming remote execution client.")
-      )
-      .arg(
         Arg::with_name("overall-deadline-secs")
             .long("overall-deadline-secs")
             .takes_value(true)
@@ -295,7 +287,6 @@ async fn main() {
     .values_of("headers")
     .map(collection_from_keyvalues::<_, BTreeMap<_, _>>)
     .unwrap_or_default();
-  let enable_streaming_client = args.is_present("enable-streaming-client");
   let overall_deadline_secs = value_t!(args.value_of("overall-deadline-secs"), u64).unwrap_or(3600);
 
   let executor = task_executor::Executor::new(Handle::current());
@@ -389,26 +380,7 @@ async fn main() {
           None
         };
 
-      let command_runner_box: Box<dyn process_execution::CommandRunner> = if enable_streaming_client
-      {
-        Box::new(
-          process_execution::remote::StreamingCommandRunner::new(
-            address,
-            ProcessMetadata {
-              instance_name: remote_instance_arg,
-              cache_key_gen_version: args.value_of("cache-key-gen-version").map(str::to_owned),
-              platform_properties,
-            },
-            root_ca_certs,
-            oauth_bearer_token,
-            headers,
-            store.clone(),
-            Platform::Linux,
-            Duration::from_secs(overall_deadline_secs),
-          )
-          .expect("Failed to make command runner"),
-        )
-      } else {
+      let command_runner_box: Box<dyn process_execution::CommandRunner> = {
         Box::new(
           process_execution::remote::CommandRunner::new(
             address,
@@ -422,10 +394,7 @@ async fn main() {
             headers,
             store.clone(),
             Platform::Linux,
-            executor,
-            std::time::Duration::from_secs(320),
-            std::time::Duration::from_millis(500),
-            std::time::Duration::from_secs(5),
+            Duration::from_secs(overall_deadline_secs),
           )
           .expect("Failed to make command runner"),
         )
